Glf dev & whales

+++ whalesdb/test/common_views.py
@@ -32,7 +32,7 @@ class CommonFormTest(TestCase):
 # Common Test for all views, this includes checking that a view is accessible or provides
 # a redirect if permissions are required to access a view
 ###########################################################################################
-class CommonTest(TestCase):
+class CommonTest(object):
     fixtures = ['initial_whale_data.json']
 
     test_url = None
@@ -75,6 +75,21 @@ class CommonTest(TestCase):
 
         return user
 
+    # Get context will use the default Test url set up in the setUp function, unless otherwise supplied as an
+    # argument here. By default this method will also log a user in as the standard whale user, unless
+    # whale_user=False, in which case a user will be logged in as a regular user not in the whale_admin group
+    def get_context(self, url=None, whale_user=True):
+        activate('en')
+
+        if whale_user:
+            self.login_whale_user()
+        else:
+            self.login_regular_user()
+
+        response = self.client.get(self.test_url if url is None else url)
+
+        return response
+
     # All views should at a minimum have a title field and determine if a user is authorized,
     # and if content is editable
     def assert_context_fields(self, response):
@@ -106,22 +121,37 @@ class CommonTest(TestCase):
 ###########################################################################################
 class CommonListTest(CommonTest):
 
+    login_required = False
+
     def setUp(self):
         super().setUp()
 
         self.test_expected_template = 'shared_models/shared_filter.html'
+
+    # login required
+    def test_view_en(self):
+        if self.login_required:
+            super().assert_view(expected_code=302)
+
+        self.login_whale_user()
+        super().assert_view()
+
+    # login required
+    def test_view_fr(self):
+        if self.login_required:
+            super().assert_view(lang='fr', expected_code=302)
+
+        self.login_whale_user()
+        super().assert_view(lang='fr')
 
     # List context should return:
     #   - a title to display in the html template
     #   - a list of fields to display
     #   - a url to use for the create button
     #   - a url to use for the detail links
-    def assert_list_view_context_fields(self):
-        activate('en')
+    def test_list_view_context_fields(self):
+        response = super().get_context()
 
-        response = self.client.get(self.test_url)
-
-        super().assert_context_fields(response)
         self.assertIn("fields", response.context)
         self.assertIn("create_url", response.context)
         self.assertIn("details_url", response.context)
@@ -148,8 +178,16 @@ class CommonCreateTest(CommonTest):
         # CreateViews intended to be used from a views.ListCommon should use the shared_entry_form.html template
         self.test_expected_template = 'whalesdb/shared_entry_form.html'
 
+    # Users must be logged in to create new objects
+    def test_view_en(self):
+        super().assert_view(expected_code=302)
+
+    # Users must be logged in to create new objects
+    def test_view_fr(self):
+        super().assert_view(lang='fr', expected_code=302)
+
     # If a user is logged in and not in 'whalesdb_admin' they should be get a 403 restriction
-    def assert_logged_in_not_access(self):
+    def test_logged_in_not_access(self):
         regular_user = self.login_regular_user()
 
         self.assertEqual(int(self.client.session['_auth_user_id']), regular_user.pk)
@@ -157,7 +195,7 @@ class CommonCreateTest(CommonTest):
         super().assert_view(expected_code=403)
 
     # If a user is logged in and in 'whalesdb_admin' they should not be redirected
-    def assert_logged_in_has_access(self):
+    def test_logged_in_has_access(self):
         whale_user = self.login_whale_user()
 
         self.assertEqual(int(self.client.session['_auth_user_id']), whale_user.pk)
@@ -165,7 +203,7 @@ class CommonCreateTest(CommonTest):
         super().assert_view()
 
     # check that the creation view is using the correct form
-    def assert_create_form(self):
+    def test_form(self):
         activate("en")
 
         view = self.expected_view
@@ -174,15 +212,13 @@ class CommonCreateTest(CommonTest):
 
     # All CommonCreate views should at a minimum have a title.
     # This will return the response for other create view tests to run further tests on context if required
-    def assert_create_view_context_fields(self):
+    def test_view_context_fields(self):
         activate('en')
 
-        self.login_whale_user()
-        response = self.client.get(self.test_url)
+        super().assert_context_fields(self.get_context())
 
-        super().assert_context_fields(response)
-
-        return response
+    def test_successful_url(self):
+        self.assert_successful_url()
 
     # test that upon a successful form the view redirects to the expected success url
     #   - Requires: self.test_url
@@ -231,14 +267,24 @@ class CommonDetailsTest(CommonTest):
     def tearDown(self) -> None:
         _details_dict = self.createDict()
 
-        for key in self._details_dict:
-            _details_dict[key].delete()
+        if self._details_dict:
+            for key in self._details_dict:
+                _details_dict[key].delete()
+
+    def test_details_en(self):
+        super().assert_view()
+
+    # Station Details are visible to all
+    def test_details_fr(self):
+        super().assert_view(lang='fr')
 
     def assert_field_in_fields(self, response):
         for field in self.fields:
             self.assertIn(field, response.context['fields'])
 
-    def assert_context_fields(self, response):
+    def test_context_fields(self):
+        response = super().get_context()
+
         super().assert_context_fields(response)
 
         self.assertIn("list_url", response.context)
